Отладка продолжается
