+++ Metrics/MetricsManager/Services/AgentsPoolRepository.cs
@@ -24,7 +24,7 @@ namespace MetricsManager.Services
             using var connection = new SQLiteConnection(databaseOptions.ConnectionString);
 
             connection.Execute(
-                "INSERT INTO agentsrepo(id, agentaddress, enable) VALUES(@id, @agentaddress, @enable)",
+                "INSERT INTO agentsrepo(agentid, agentaddress, enable) VALUES(@id, @agentaddress, @enable)",
                 new
                 {
                     id = item.AgentId,   // не могу понять, почему этот параметр в базе равен 0
@@ -38,7 +38,7 @@ namespace MetricsManager.Services
             using var connection = new SQLiteConnection(_databaseOptions.Value.ConnectionString);
 
             connection.Execute(
-                "DELETE FROM agentsrepo WHERE Id=@id",
+                "DELETE FROM agentsrepo WHERE agentid=@id",
                 new
                 {
                     id = id
@@ -50,7 +50,7 @@ namespace MetricsManager.Services
             using var connection = new SQLiteConnection(_databaseOptions.Value.ConnectionString);
 
             connection.Execute(
-                "UPDATE agentsrepo SET agentaddress = @agentaddress, enable = @enable WHERE Id = @id",
+                "UPDATE agentsrepo SET agentaddress = @agentaddress, enable = @enable WHERE agentid = @id",
                 new
                 {
                     agentaddress = item.AgentAddress.ToString(),
@@ -63,13 +63,8 @@ namespace MetricsManager.Services
         {
             using var connection = new SQLiteConnection(_databaseOptions.Value.ConnectionString);
 
-            //List<AgentInfoDto> agentsq = connection.Query<AgentInfoDto>(
-            //    "SELECT * FROM agentsrepo").ToList();
-
-            //Dictionary<int, AgentInfoDto> agents = agentsq.ToDictionary(x => x.AgentId, x => x);
-
             Dictionary<int, AgentInfoDto> agents = connection.Query<AgentInfoDto>(
-                "SELECT * FROM agentsrepo").ToDictionary(x => x.AgentId, y => y); // здесь соответственно возникает исключение если более одной записи, т.к. значения колонки id равны нулю
+                "SELECT * FROM agentsrepo").ToDictionary(x => x.AgentId, y => y); 
 
             return agents;
         }
@@ -79,7 +74,7 @@ namespace MetricsManager.Services
             using var connection = new SQLiteConnection(_databaseOptions.Value.ConnectionString);
 
             AgentInfoDto agent = connection.QuerySingle<AgentInfoDto>(
-                "SELECT Id, agentaddress, enable FROM agentsrepo WHERE Id = @id",
+                "SELECT agentid, agentaddress, enable FROM agentsrepo WHERE agentid = @id",
                 new
                 {
                     id = id
